Separating styles into differnet file in introScreen:
