+++ givantk-mobile/src/screens/commons/IntroScreen/IntroScreenStyles.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    mainContent: {
+      flex: 1,
+    },
+    image: {
+      width: 320,
+      height: 320,
+    },
+    text: {
+      color: 'rgba(255, 255, 255, 0.8)',
+      backgroundColor: 'transparent',
+      textAlign: 'center',
+      paddingHorizontal: 18,
+    },
+    title: {
+      fontSize: 22,
+      color: 'white',
+      backgroundColor: 'transparent',
+      textAlign: 'center',
+      marginBottom: 16,
+    },
+  });
+export default styles;
